refactor src-tauri/
